+++ backend/src/main/java/com/jackpot/narratix/domain/repository/CoverLetterJpaRepository.java
@@ -24,5 +24,5 @@ public interface CoverLetterJpaRepository extends JpaRepository<CoverLetter, Lon
             String userId, LocalDate startDate, LocalDate endDate, Pageable pageable
     );
 
-    Integer countByUserIdAndDeadlineBetween(String userId, LocalDate startDate, LocalDate endDate);
+    Long countByUserIdAndDeadlineBetween(String userId, LocalDate startDate, LocalDate endDate);
 }
